temp, completed tracking is failing, need to set complete on done
refactor steps schema

turn into components, refactor steps

add note on completion tracker
