+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/seo-assistant/use-keywords-step.tsx
@@ -1,26 +1,23 @@
 import { createInterpolateElement, useCallback, useState } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
-import type { Step } from './index';
+import type { Step } from './types';
 
-export const useKeywordsStep = ( {
-	addMessage,
-	onStep,
-} ): {
-	stepProps: Step;
-	value: string;
-	setValue: React.Dispatch< React.SetStateAction< string > >;
-} => {
+export const useKeywordsStep = ( { addMessage, onStep } ): Step => {
 	const [ keywords, setKeywords ] = useState( '' );
+	const [ completed, setCompleted ] = useState( false );
 
 	const handleSkip = useCallback( () => {
-		addMessage( __( 'Ok, skipping for now.', 'jetpack' ) );
-	}, [ addMessage ] );
+		addMessage( __( 'Ok, skipping keywords.', 'jetpack' ) );
+		if ( onStep ) {
+			onStep( { value: '' } );
+		}
+	}, [ addMessage, onStep ] );
 
 	const handleKeywordsSubmit = useCallback( () => {
-		addMessage( { content: keywords, isUser: true } );
-		if ( ! keywords ) {
+		if ( ! keywords.trim() ) {
 			return handleSkip();
 		}
+		addMessage( { content: keywords, isUser: true } );
 
 		const keywordlist = keywords
 			.split( ',' )
@@ -42,45 +39,48 @@ export const useKeywordsStep = ( {
 			}
 		);
 		addMessage( { content: message } );
+		setCompleted( true );
 		if ( onStep ) {
 			onStep( { value: keywords } );
 		}
 	}, [ onStep, addMessage, keywords, handleSkip ] );
 
 	return {
-		stepProps: {
-			id: 'keywords',
-			title: __( 'Optimise for SEO', 'jetpack' ),
-			messages: [
-				{
-					content: createInterpolateElement(
-						__( "<b>Hi there! 👋 Let's optimise your blog post for SEO.</b>", 'jetpack' ),
-						{ b: <b /> }
-					),
-					showIcon: true,
-				},
-				{
-					content: createInterpolateElement(
-						__(
-							"Here's what we can improve:<br />1. Keywords<br />2. Title<br />3. Meta description",
-							'jetpack'
-						),
-						{ br: <br /> }
-					),
-					showIcon: false,
-				},
-				{
-					content: __(
-						'To start, please enter 1–3 focus keywords that describe your blog post.',
+		id: 'keywords',
+		title: __( 'Optimise for SEO', 'jetpack' ),
+		label: __( 'Keywords', 'jetpack' ),
+		messages: [
+			{
+				content: createInterpolateElement(
+					__( "<b>Hi there! 👋 Let's optimise your blog post for SEO.</b>", 'jetpack' ),
+					{ b: <b /> }
+				),
+				showIcon: true,
+			},
+			{
+				content: createInterpolateElement(
+					__(
+						"Here's what we can improve:<br />1. Keywords<br />2. Title<br />3. Meta description",
 						'jetpack'
 					),
-					showIcon: true,
-				},
-			],
-			type: 'input',
-			placeholder: __( 'Photography, plants', 'jetpack' ),
-			onSubmit: handleKeywordsSubmit,
-		},
+					{ br: <br /> }
+				),
+				showIcon: false,
+			},
+			{
+				content: __(
+					'To start, please enter 1–3 focus keywords that describe your blog post.',
+					'jetpack'
+				),
+				showIcon: true,
+			},
+		],
+		type: 'input',
+		placeholder: __( 'Photography, plants', 'jetpack' ),
+		onSubmit: handleKeywordsSubmit,
+		onSkip: handleSkip,
+		completed,
+		setCompleted,
 		value: keywords,
 		setValue: setKeywords,
 	};
